pi0disp_bak/pyproject.toml
[project]
name = "pi0disp"
version = "0.1.0"
description = "A driver for the ST7789V display on Raspberry Pi."
authors = [
    {name = "Nil Niu", email = "nil.ni.creator@gmail.com"},
]
dependencies = [
    "click",
    "numpy",
    "pigpio",
    "pillow",
    "ninja_utils",
]
requires-python = ">=3.9"
readme = "README.md"
license = {file = "LICENSE"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
pi0disp = "pi0disp.__main__:cli"
